+++ e2e/options/sections-tests.js
@@ -1,4 +1,4 @@
-import { test, expect, timeout } from "../fixtures.js";
+import { test, timeout } from "../fixtures.js";
 import { OptionsPage } from "../pom/index.js";
 
 test.describe("Sections", () => {
@@ -11,17 +11,27 @@ test.describe("Sections", () => {
         await context.pages()[0].close();
     });
 
-    test("Sections", async ({ page }) => {
-        // Arrange
-        const sections = page.locator("h2");
-
-        // Assert
-        await expect(sections.nth(0)).toHaveText("Autoclose");
-    });
-
-    test("Section: Autoclose", async ({ page }) => {
+    test("Appearance", async ({ page }) => {
         // Arrange
         const pom = new OptionsPage(page);
+
+        await pom.ui.selectedItemColor.hasValue("#EC4339");
+
+        // Act
+        await pom.ui.selectedItemColor.setValue("#00A0DC");
+
+        await pom.save();
+        await pom.reload();
+
+        // Assert
+        await pom.ui.selectedItemColor.hasValue("#00A0DC");
+    });
+
+    test("Autoclose", async ({ page }) => {
+        // Arrange
+        const pom = new OptionsPage(page);
+
+        await pom.getPin(2).click();
 
         await pom.autoclose.enabled.isChecked(true);
         await pom.autoclose.time.hasValue("1");