+++ e2e/options/tabs-tests.js
@@ -9,16 +9,68 @@ test.describe("Tabs", () => {
         await pom.goto();
     });
 
-    test("Titles and descriptions", async ({ page }) => {
-        // Assert
-        const titles = await page.locator("span.title").allInnerTexts();
-        for (const title of titles) {
-            await expect(title.length).toBeGreaterThan(0);
-        }
+    const COUNT = 2;
 
-        const icons = await page.locator("span.info").allInnerTexts();
-        for (const icon of icons) {
-            await expect(icon.length).toBe(5);
+    test("Pins", async ({ page }) => {
+        // Arrange
+        const pom = new OptionsPage(page);
+
+        // Assert
+        await expect(pom.getPin(1)).toHaveText("Appearance");
+        await expect(pom.getPin(2)).toHaveText("Autoclose");
+    });
+
+    test("Tabs", async ({ page }) => {
+        // Arrange
+        const pom = new OptionsPage(page);
+
+        // Assert
+        await expect(pom.getTab(1).locator("h2")).toHaveText("Appearance");
+        await expect(pom.getTab(2).locator("h2")).toHaveText("Autoclose");
+
+        for (let i = 1; i <= COUNT; i++) {
+            if (i === 1) {
+                await expect(pom.getTab(i)).toBeVisible();
+            } else {
+                await expect(pom.getTab(i)).toBeHidden();
+            }
+        }
+    });
+
+    test("Tabs visibility", async ({ page }) => {
+        // Arrange
+        const pom = new OptionsPage(page);
+
+        // Assert
+        for (let i = 1; i <= COUNT; i++) {
+            await pom.getPin(i).click();
+            for (let j = 1; j <= COUNT; j++) {
+                if (i === j) {
+                    await expect(pom.getTab(j)).toBeVisible();
+                } else {
+                    await expect(pom.getTab(j)).toBeHidden();
+                }
+            }
+        }
+    });
+
+    test("Titles and descriptions", async ({ page }) => {
+        // Arrange
+        const pom = new OptionsPage(page);
+
+        // Assert
+        for (let i = 1; i <= COUNT; i++) {
+            await pom.getPin(i).click();
+
+            const titles = await pom.getTab(i).locator("span.title").allInnerTexts();
+            for (const title of titles) {
+                await expect(title.length).toBeGreaterThan(0);
+            }
+
+            const icons = await pom.getTab(i).locator("span.info").allInnerTexts();
+            for (const icon of icons) {
+                await expect(icon.length).toBe(5);
+            }
         }
     });
 
@@ -30,19 +82,23 @@ test.describe("Tabs", () => {
         await modal.hidden();
 
         // Assert
-        const icons = await page.locator("span.info").all();
-        for (const icon of icons) {
-            await icon.click();
-            await modal.visible();
-            await modal.hasTitle();
-            await modal.hasDescription();
-            await modal.close();
-            await modal.hidden();
+        for (let i = 1; i <= COUNT; i++) {
+            await pom.getPin(i).click();
 
-            await icon.click();
-            await modal.visible();
-            await modal.exit();
-            await modal.hidden();
+            const icons = await pom.getTab(i).locator("span.info").all();
+            for (const icon of icons) {
+                await icon.click();
+                await modal.visible();
+                await modal.hasTitle();
+                await modal.hasDescription();
+                await modal.close();
+                await modal.hidden();
+
+                await icon.click();
+                await modal.visible();
+                await modal.exit();
+                await modal.hidden();
+            }
         }
     });
 });