+++ src/common/constants/defaults.js
@@ -1,7 +1,10 @@
 import * as OPTIONS from "./options.js";
 import { T1 } from "./autoclose.js";
+import { RED } from "./colors.js";
 
 export const DEFAULTS = {
     [OPTIONS.IS_AUTOCLOSE_ENABLED]: true,
-    [OPTIONS.AUTOCLOSE_TIME]: T1
+    [OPTIONS.AUTOCLOSE_TIME]: T1,
+
+    [OPTIONS.UI_SELECTED_ITEM_COLOR]: RED
 };