+++ src/common/constants/index.js
@@ -4,3 +4,4 @@ export * as OPTIONS from "./options.js";
 export * as TARGETS from "./targets.js";
 export * as MENU from "./menu.js";
 export * as AUTOCLOSE from "./autoclose.js";
+export * as COLORS from "./colors.js";