+++ src/common/constants/options.js
@@ -1,2 +1,4 @@
 export const IS_AUTOCLOSE_ENABLED = "is-autoclose-enabled";
 export const AUTOCLOSE_TIME = "autoclose-time";
+
+export const UI_SELECTED_ITEM_COLOR = "ui-selected-item-color";